requirements.txt
# Python dependencies for TaskFlow AI Customer Success Agent

# Core dependencies
anthropic>=0.40.0
openai>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0

# Database
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.23

# Message queue (Stage 2)
kafka-python>=2.0.2

# Channel integrations
google-api-python-client>=2.108.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
twilio>=8.10.0

# NLP and sentiment analysis
transformers>=4.35.0
torch>=2.1.0
nltk>=3.8.1

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
aiohttp>=3.9.0
apscheduler>=3.10.0

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0

# Development
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0
